+++ crates/ra_macros/src/lib.rs
@@ -12,3 +12,4 @@ macro_rules! impl_froms {
 
 pub mod tt;
 pub mod mbe;
+mod mbe_parser;
